Modularize folders to separate cell and board functionality
