+++ src/board/data/type.ts
@@ -1,0 +1,3 @@
+import { Cell } from "../../cell/data/type.js";
+
+export type Board = Cell[][];
