+++ src/board/factory/generateBoard.ts
@@ -1,5 +1,5 @@
 import { Board } from "../data/type.js";
-import { Cell } from "../data/type.js";
+import { Cell } from "../../cell/data/type.js";
 
 export const generateBoard = (boardDimension: number): Board => {
   if (boardDimension < 5) {
