+++ src/index.ts
@@ -1,4 +1,4 @@
-import { generateBoard } from "./board/generateBoard.js";
+import { generateBoard } from "./board/factory/generateBoard.js";
 import { generateBoardWithMines } from "./board/generateBoardWithMines.js";
 import { generateCompleteBoard } from "./board/generateCompleteBoard.js";
 import {
@@ -6,7 +6,7 @@ import {
   getAdjacentMinesTotal,
 } from "./board/getAdjacentMines.js";
 import { getRandomPositions } from "./board/getRandomPositions.js";
-import { renderBoard } from "./ui/renderBoard.js";
+import { renderBoard } from "./board/ui/renderBoard.js";
 
 export const board = generateBoard(8);
